fix: move out handler.rs tests
